requirements.txt
aiohttp==2.3.10
cchardet==1.0.0
gunicorn==19.3.0
httpie==0.9.2
PyJWT==1.4.0
